requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
requests==2.31.0
groq==0.4.1
pymupdf==1.23.7
python-docx==1.0.1
python-dotenv==1.0.0
pydantic==2.5.2
typing-extensions==4.8.0
httpx==0.25.0
